# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Runic API
  description: |
    Firewall rule management system API.
    
    ## Authentication
    Most endpoints require Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Error Responses
    All endpoints may return error responses in the following format:
    ```json
    {
      "error": "error message"
    }
    ```
    
    Common HTTP status codes:
    - 400: Bad Request - Invalid input
    - 401: Unauthorized - Missing or invalid token
    - 403: Forbidden - Insufficient permissions
    - 404: Not Found - Resource not found
    - 409: Conflict - Resource in use
    - 500: Internal Server Error
  version: 1.0.0
  contact:
    name: Runic Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: /api/v1
    description: Production server (relative path)
...
paths:
...
  /peers:
    get:
      tags:
        - Peers
      summary: List all peers
      description: Returns a list of all peers with their status, groups, and metadata.
      operationId: listPeers
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Peer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    post:
      tags:
        - Peers
      summary: Create a new peer
      description: |
        Creates a new peer (agent or manual).
        
        For agent peers, `agent_key` is required.
        For manual peers, `hostname` and `ip_address` are required.
      operationId: createPeer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePeerRequest'
      responses:
        '201':
          description: Peer created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 123
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
...
  /groups:
    get:
      tags:
        - Groups
      summary: List all groups
      description: Returns a list of all groups with peer and policy counts.
      operationId: listGroups
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    post:
      tags:
        - Groups
      summary: Create a new group
      description: Creates a new group.
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Group created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
...
  /services:
    get:
      tags:
        - Services
      summary: List all services
      description: Returns a list of all services.
      operationId: listServices
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    post:
      tags:
        - Services
      summary: Create a new service
      description: Creates a new service.
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceRequest'
      responses:
        '201':
          description: Service created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
...
  /policies:
    get:
      tags:
        - Policies
      summary: List all policies
      description: Returns a list of all policies.
      operationId: listPolicies
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Policy'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
    post:
      tags:
        - Policies
      summary: Create a new policy
      description: Creates a new firewall policy.
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePolicyRequest'
      responses:
        '201':
          description: Policy created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    format: int64
                    example: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - BearerAuth: []
...
  /setup:
    get:
      tags:
      - Auth
      summary: Get setup status
      description: |
        Returns whether initial setup has been completed.
        This endpoint does not require authentication.
      operationId: getSetupStatus
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  is_setup:
                    type: boolean
                    example: false
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
      - Auth
      summary: Complete initial setup
      description: |
        Completes the initial setup by creating the admin user and JWT secret.
        This endpoint does not require authentication and can only be called once.
      operationId: completeSetup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
      responses:
        '200':
          description: Setup completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "setup complete"
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags:
        - Auth
      summary: User login
      description: Authenticates user and returns access/refresh tokens.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "runic_access_token=...; HttpOnly; Path=/"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "login successful"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: User logout
      description: Clears authentication cookies and invalidates session.
      operationId: logout
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "logged out"
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh access token
      description: |
        Refreshes the access token using a valid refresh token.
        The refresh token is read from cookies.
      operationId: refreshToken
      responses:
        '200':
          description: Token refreshed successfully
          headers:
            Set-Cookie:
              schema:
                type: string
                example: "runic_access_token=...; HttpOnly; Path=/"
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "token refreshed"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /dashboard:
    get:
      tags:
      - Dashboard
      summary: Get dashboard data
      description: Returns system-wide statistics, peer health, and recent activity for the dashboard.
      operationId: getDashboard
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /logs:
    get:
      tags:
      - Logs
      summary: Get firewall logs
      description: Returns firewall logs with optional filtering.
      operationId: getLogs
      parameters:
      - name: limit
        in: query
        schema:
          type: integer
          default: 100
      - name: offset
        in: query
        schema:
          type: integer
          default: 0
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FirewallLog'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /logs/stream:
    get:
      tags:
      - Logs
      summary: Stream firewall logs
      description: |
        Opens a WebSocket connection to stream real-time firewall log events.
        Upgrades the HTTP connection to WebSocket protocol.
      operationId: streamLogs
      responses:
        '101':
          description: WebSocket upgrade successful
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - BearerAuth: []

  # ==================== PENDING CHANGES ====================
  /pending-changes:
    get:
      tags:
      - Pending Changes
      summary: Get pending changes
      description: Returns all pending changes grouped by peer.
      operationId: listPendingChanges
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingChangeGroup'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /pending-changes/{peerId}:
    get:
      tags:
      - Pending Changes
      summary: Get peer pending changes
      description: Returns all pending changes for a specific peer.
      operationId: getPeerPendingChanges
      parameters:
      - name: peerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingChangeDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /pending-changes/{peerId}/preview:
    post:
      tags:
      - Pending Changes
      summary: Preview pending bundle
      description: Previews the compiled iptables rules from pending changes for a peer. Requires editor role.
      operationId: previewPeerPendingBundle
      parameters:
      - name: peerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Preview result
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                    description: The compiled iptables rules preview
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /pending-changes/{peerId}/apply:
    post:
      tags:
      - Pending Changes
      summary: Apply peer pending bundle
      description: Applies all pending changes for a specific peer. Requires editor role.
      operationId: applyPeerPendingBundle
      parameters:
      - name: peerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: Bundle applied successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "bundle applied"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
...
  /pending-changes/rollback:
    post:
      tags:
      - Pending Changes
      summary: Rollback pending changes
      description: Rolls back all pending changes. Requires editor role.
      operationId: rollbackPendingChanges
      responses:
        '200':
          description: Changes rolled back
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "changes rolled back"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
  /pending-changes/apply-all:
    post:
      tags:
      - Pending Changes
      summary: Apply all pending bundles
      description: Applies all pending changes across all peers. Requires editor role.
      operationId: applyAllPendingBundles
      responses:
        '200':
          description: All bundles applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "all bundles applied"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        is_system:
          type: boolean
        peer_count:
          type: integer
        policy_count:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Bad Request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'